% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_active_events.R
\name{list_active_events}
\alias{list_active_events}
\title{List active events}
\usage{
list_active_events(api)
}
\arguments{
\item{api}{An API key from Sportity.}
}
\value{
A data frame.
}
\description{
List active events on Sportity.
}
\examples{
\dontrun{
list_active_events(api = "your_api_key")
}
}
