% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_documents.R
\name{list_documents}
\alias{list_documents}
\title{List documents}
\usage{
list_documents(api, password, event_id = "")
}
\arguments{
\item{api}{An API key from Sportity.}

\item{password}{The event/series password.}

\item{event_id}{The id of the event. To list all document in a series do not
specify an ´event_id´. If to only list documents in a particular event, then
specify ´event_id´.}
}
\value{
A data frame.
}
\description{
List all documents in an event or series.
}
\examples{
\dontrun{
list_documents(
  api = "your_api_key",
  password = "your_password"
)
}
}
