% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_past_events.R
\name{list_past_events}
\alias{list_past_events}
\title{List past events}
\usage{
list_past_events(api)
}
\arguments{
\item{api}{An API key from Sportity.}
}
\value{
A data frame.
}
\description{
List past events on Sportity.
}
\examples{
\dontrun{
list_past_events(api = "your_api_key")
}
}
