% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_document.R
\name{reorder_document}
\alias{reorder_document}
\title{Reorder document}
\usage{
reorder_document(api, document_id, position)
}
\arguments{
\item{api}{An API key from Sportity}

\item{document_id}{The id of the document to be reordered. The id can either
be retrieved using ´list_documents´ or from Sportity.}

\item{position}{The new position of the document.}
}
\description{
Reorder a document in an event or series.
}
\examples{
\dontrun{
reorder_document(
  api = "your_api_key",
  document_id = "your_document_id",
  position = 1
)
}
}
