% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_folder.R
\name{create_folder}
\alias{create_folder}
\title{Create folder document}
\usage{
create_folder(api, password, document_name, event_id = NULL)
}
\arguments{
\item{api}{An API key from Sportity.}

\item{password}{The event/series password.}

\item{document_name}{The name of the folder to be created.}

\item{event_id}{The id of the event. To create series folder do not specify an ´event_id´. If to create a document in a particular event, then specify ´event_id´.}
}
\description{
Create a folder document in an event or series.
}
