% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_image.R
\name{create_image}
\alias{create_image}
\title{Create image document}
\usage{
create_image(api, password, document_name, filepath, event_id = NULL)
}
\arguments{
\item{api}{An API key from Sportity.}

\item{password}{The event/series password.}

\item{document_name}{The name of the document to be created.}

\item{event_id}{The id of the event. To create series folder do not specify an ´event_id´. If to create a document in a particular event, then specify ´event_id´.}
}
\description{
Create an image document in an event or series.
}
