% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_link.R
\name{create_link}
\alias{create_link}
\title{Create link document}
\usage{
create_link(api, password, document_name, url, event_id = NULL)
}
\arguments{
\item{api}{An API key from Sportity.}

\item{password}{The event/series password.}

\item{document_name}{The name of the document to be created.}

\item{url}{The link to create a link document from.}

\item{event_id}{The id of the event. To create series folder do not specify an ´event_id´. If to create a document in a particular event, then specify ´event_id´.}
}
\description{
Create a link document in an event or series.
}
