% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pdf.R
\name{create_pdf}
\alias{create_pdf}
\title{Create/upload pdf}
\usage{
create_pdf(
  api,
  password,
  filepath = file.choose(),
  document_name = tools::file_path_sans_ext(basename(filepath)),
  event_id = NULL
)
}
\arguments{
\item{api}{An API key from Sportity.}

\item{password}{The event/series password.}

\item{filepath}{The path to the document to be created/uploaded.}

\item{document_name}{The id of the document to be deleted.}

\item{event_id}{The id of the event. To create series folder do not specify an ´event_id´. If to create a document in a particular event, then specify ´event_id´.}
}
\description{
Create a pdf document in an event or series.
}
