% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_text.R
\name{create_text}
\alias{create_text}
\title{Create text document}
\usage{
create_text(api, password, document_name, text, event_id = NULL)
}
\arguments{
\item{api}{An API key from Sportity.}

\item{password}{The event/series password.}

\item{document_name}{The name of the document to be created.}

\item{event_id}{The id of the event. To create series folder do not specify an ´event_id´. If to create a document in a particular event, then specify ´event_id´.}
}
\description{
Create a text document in an event or series.
}
