% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_document.R
\name{delete_document}
\alias{delete_document}
\title{Delete document}
\usage{
delete_document(api, document_id)
}
\arguments{
\item{api}{An API key from Sportity}

\item{document_id}{The id of the document to be deleted. The id can either be retrieved using ´list_documents´ or from Sportity.}
}
\description{
Delete a document from an event or series.
}
