% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_all_events.R
\name{list_all_events}
\alias{list_all_events}
\title{List all events}
\usage{
list_all_events(api)
}
\arguments{
\item{api}{An API key from Sportity.}
}
\value{
A data frame.
}
\description{
List all events on Sportity - active, past and deleted.
}
