% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_deleted_events.R
\name{list_deleted_events}
\alias{list_deleted_events}
\title{List deleted events}
\usage{
list_deleted_events(api)
}
\arguments{
\item{api}{An API key from Sportity.}
}
\value{
A data frame.
}
\description{
List deleted events on Sportity
}
