% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_personal_logins.R
\name{list_personal_logins}
\alias{list_personal_logins}
\title{List personal logins}
\usage{
list_personal_logins(api, password)
}
\arguments{
\item{api}{An API key from Sportity.}

\item{password}{The event/series password.}
}
\value{
A data frame.
}
\description{
List personal logins for an event or series.
}
